demo_mes/requirements.txt
asgiref==3.5.2
Django==4.1.3
django-cors-headers==3.13.0
djangorestframework==3.14.0
mysqlclient==2.1.1
pytz==2022.6
sqlparse==0.4.3
tzdata==2022.6
gunicorn==20.0.4